requirements.txt
numpy
ntfy
slackclient
ntfy[slack]
tensorflow-gpu==1.30
tensorflow-datasets==1.10
keyboard
opencv-python
netron
wget
pathlib
statistics
argparse
scikit-optimize
matplotlib
scipy
Pillow
seaborn
